pyproject.toml
[tool.poetry]
name = "pycep"
version = "2.0.0"
description = "A simple cli that get cep information"
authors = ["Guilherme Seganzerla <g.seganzerla@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.27.1"
click = "^8.0.3"

[tool.poetry.dev-dependencies]
yapf = "^0.32.0"
pylint = "^2.12.2"
poetry2setup = "^1.0.0"
vcrpy = "^4.1.1"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
pycep = 'pycep:cep_cli'
